Add files via upload
1.State the null and the alternate hypothesis for conducting one-way ANOVA for both Education and Occupation individually.
Null hypothesis (H0) is the presumed current state of the matter or status quo. Unless there is strong evidence otherwise, null hypothesis is not rejected (HA) is the rival opinion or research hypothesis.
a. Null hypothesis related to Education
H0: There is no effect of education on salary
HA: For at least on education level, mean salary is different from the others
H0: µA = µs = µP = µE
HA: At least one mean is different from the others
2.Perform a one-way ANOVA on Salary with respect to Education. State whether the null hypothesis is accepted or rejected based on the ANOVA results.
First deciding the level of significance:
The level of significance is defined as the probability of rejecting a null hypothesis when it is true, and is denoted by o. That is, P (Type I error) = o = 0.05
The below is output from F oneway from scipy.stats
Interpretation: The P-value obtained from ANOVA for Education is statistically significant as (P 0.05). We conclude Salary depends on the levels of Education. lvtean salary for at least one level of Education is different from the others.
3. Perform a one-way ANOVA on Salary with respect to Occupation. State whether the null hypothesis is accepted or rejected based on the ANOVA results.
The below is output from F oneway from scip.y.stats.
Interpretation: The null hypothesis that mean salary is different for different level of occupation is not rejected.
4. If null hypothesis is rejected in either (1.2) or in (1.3), find out which class means are significantly different. Interpret the result.
The null hypothesis that mean salary is different for different levels of educational qualification is rejected, but the null hypothesis that mean salary is different for different level of occupation is not rejected. Hence we need to find out for which educational qualification mean salary is significantly different fr0m the other levels.